refactorizar para recuperar usuario de la request (token)

+++ src/campanas/campanas.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   Logger,
   NotFoundException,
@@ -30,10 +31,11 @@ export class CampanasService {
     private readonly campanasContactosRepository: Repository<CampanaContacto>,
   ) {}
 
-  async create(createCampanaDto: CreateCampanaDto) {
+  async create(createCampanaDto: CreateCampanaDto, usuarioId: number) {
     try {
       const campana = this.campanasRepository.create({
         estado: Estados.PENDIENTE,
+        usuarioId,
         ...createCampanaDto,
       });
 
@@ -63,7 +65,11 @@ export class CampanasService {
     return campana;
   }
 
-  async update(id: number, updateCampanaDto: UpdateCampanaDto) {
+  async update(
+    id: number,
+    updateCampanaDto: UpdateCampanaDto,
+    usuarioId: number,
+  ) {
     const campana = await this.campanasRepository.preload({
       id,
       ...updateCampanaDto,
@@ -71,6 +77,12 @@ export class CampanasService {
 
     if (!campana) {
       throw new NotFoundException(`Campaña con id ${id} no encontrado`);
+    }
+
+    if (campana.usuarioId !== usuarioId) {
+      throw new ForbiddenException(
+        'No tienes permiso para actualizar esta campaña',
+      );
     }
 
     try {
@@ -82,8 +94,14 @@ export class CampanasService {
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, usuarioId: number) {
     const campana = await this.findOne(id);
+
+    if (campana.usuarioId !== usuarioId) {
+      throw new ForbiddenException(
+        'No tienes permiso para eliminar esta campaña',
+      );
+    }
 
     try {
       await this.campanasRepository.remove(campana);
@@ -100,7 +118,7 @@ export class CampanasService {
     throw new BadRequestException(`Error: ${error.message}`);
   }
 
-  async iniciar(id: number) {
+  async iniciar(id: number, usuarioId: number) {
     const campana = await this.campanasRepository.findOne({
       where: { id: Equal(id) },
       relations: {
@@ -110,6 +128,12 @@ export class CampanasService {
 
     if (!campana) {
       throw new NotFoundException(`Campaña con id ${id} no encontrado`);
+    }
+
+    if (campana.usuarioId !== usuarioId) {
+      throw new ForbiddenException(
+        'No tienes permiso para iniciar esta campaña',
+      );
     }
 
     if (campana.etiquetas.length === 0) {
@@ -159,7 +183,7 @@ export class CampanasService {
     }
   }
 
-  async cancelar(id: number) {
+  async cancelar(id: number, usuarioId: number) {
     const campana = await this.campanasRepository.findOne({
       where: { id: Equal(id) },
       relations: {
@@ -169,6 +193,12 @@ export class CampanasService {
 
     if (!campana) {
       throw new NotFoundException(`Campaña con id ${id} no encontrado`);
+    }
+
+    if (campana.usuarioId !== usuarioId) {
+      throw new ForbiddenException(
+        'No tienes permiso para cancelar esta campaña',
+      );
     }
 
     if (campana.estado !== Estados.EN_PROCESO) {
@@ -186,7 +216,7 @@ export class CampanasService {
     }
   }
 
-  async finalizar(id: number) {
+  async finalizar(id: number, usuarioId: number) {
     const campana = await this.campanasRepository.findOne({
       where: { id: Equal(id) },
       relations: {
@@ -196,6 +226,12 @@ export class CampanasService {
 
     if (!campana) {
       throw new NotFoundException(`Campaña con id ${id} no encontrado`);
+    }
+
+    if (campana.usuarioId !== usuarioId) {
+      throw new ForbiddenException(
+        'No tienes permiso para finalizar esta campaña',
+      );
     }
 
     if (campana.estado !== Estados.EN_PROCESO) {
